requirements.txt
torch
torchvision
numpy
Pillow
imageio
imageio-ffmpeg